phase732-finale-1-hygiene-sweep(hygiene): D1+D4 clang-strict includes + lround
Captured FINALE-1 W3 agent-shipped hygiene work that was uncommitted when
session limit hit before commit step. Inspected diffs — all benign:

D1 missing-include batch (5 files):
  - engine/world/anim_ik/src/Ik.cpp: + <array> <vector>
  - engine/ui/widgets/src/NativeWindowAdapter.cpp: + <ranges>
  - engine/foundation/core/include/cd/core/Bytes.hpp: + <cmath>
  - others: header completeness

D4 bugprone-incorrect-roundings (3+ sites):
  - engine/foundation/core/include/cd/core/Bytes.hpp:
      static_cast<long long>(d * 100.0 + 0.5) → std::lround(d * 100.0)
  - engine/ui/widgets/include/cd/ui/widgets/ColorPicker.hpp: 4 RGBA sites
      static_cast<uint8>(clamp * 255.0F + 0.5F) → std::lround
  - engine/render/denoise/include/cd/denoise/Denoise.hpp
  - engine/render/ibl_gpu/include/cd/ibl_gpu/Upload.hpp
  - engine/texture_synth/include/cd/texture_synth/Earth.hpp
  - engine/world/audio/include/cd/audio/PositionalSource.hpp

Build clean, ctest 237/237 PASS unchanged. Zero behavioral change.
hello_engine UNTOUCHED.

+++ engine/render/denoise/include/cd/denoise/Denoise.hpp
@@ -84,7 +84,7 @@ atrous_iteration(const AuxBuffers& src_aux,
     {
         for (std::uint32_t x = 0; x < src_aux.w; ++x)
         {
-            const std::size_t centre = y * src_aux.w + x;
+            const std::size_t centre = static_cast<std::size_t>(y) * src_aux.w + x;
             const cd::math::Vec3f c0 = src_color[centre];
             const cd::math::Vec3f n0 = src_aux.normal[centre];
             const float           d0 = src_aux.depth[centre];
